Time: 0 ms (100.00%) | Memory: 6.9 MB (17.43%) - LeetSync

+++ 392-is-subsequence/is-subsequence.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isSubsequence(string s, string t) 
     {
-        int c=0;
+        short int c=0;
         for(int i=0;t[i]!='\0';i++)
             if(t[i]==s[c])
             ++c;
